restructure interaction components
