+++ frontend/src/modules/interactions/components/InteractionHistory/InteractionHistory.module.scss
