+++ frontend/src/modules/interactions/components/InteractionIcon/InteractionIcon.module.scss
