+++ frontend/src/modules/interactions/components/InteractionIcon/InteractionIcon.tsx
