+++ frontend/src/modules/interactions/components/InteractionLabel/InteractionLabel.module.scss
