+++ frontend/src/modules/interactions/components/InteractionOutcome/InteractionOutcome.module.scss
