+++ frontend/src/modules/interactions/components/InteractionOutcome/InteractionOutcome.tsx
