+++ frontend/src/modules/interactions/components/ValueBuilder/ArrayBuilder/ArrayBuilder.tsx
@@ -3,8 +3,8 @@ import { CadenceValueBuilder } from "../interface";
 import { ValueBuilder } from "../ValueBuilder";
 import classes from "./ArrayBuilder.module.scss";
 import { FclArrayValue, FclValue, FclValues } from "@flowser/shared";
-import { SimpleButton } from "../../../../../../components/buttons/simple-button/SimpleButton";
-import { SizedBox } from "../../../../../../components/sized-box/SizedBox";
+import { SimpleButton } from "../../../../../components/buttons/simple-button/SimpleButton";
+import { SizedBox } from "../../../../../components/sized-box/SizedBox";
 
 export function ArrayBuilder(props: CadenceValueBuilder): ReactElement {
   const { type, value, setValue } = props;
