+++ frontend/src/modules/interactions/components/ValueBuilder/DictionaryBuilder/DictionaryBuilder.module.scss
