+++ frontend/src/modules/interactions/components/ValueBuilder/FixedPointNumberBuilder/FixedPointNumberBuilder.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement, useEffect } from "react";
 import { CadenceValueBuilder } from "../interface";
-import Input from "../../../../../../components/inputs/input/Input";
+import Input from "../../../../../components/inputs/input/Input";
 import { FclValues } from "@flowser/shared";
 
 export function FixedPointNumberBuilder(
