+++ frontend/src/modules/interactions/components/ValueBuilder/IntegerNumberBuilder/IntegerNumberBuilder.tsx
@@ -1,17 +1,18 @@
 import React, { ReactElement, useEffect } from "react";
 import { CadenceValueBuilder } from "../interface";
-import Input from "../../../../../../components/inputs/input/Input";
+import Input from "../../../../../components/inputs/input/Input";
+import { FclValues } from "@flowser/shared";
 
-export function TextualBuilder(props: CadenceValueBuilder): ReactElement {
+export function IntegerNumberBuilder(props: CadenceValueBuilder): ReactElement {
   const { value, setValue } = props;
 
-  const isInitialised = typeof value === "string";
+  const isInitialised = FclValues.isFclIntegerNumberValue(value);
 
   // TODO(polish): Don't trigger this hook on every rerender
   //  See: https://www.notion.so/flowser/Sometimes-arguments-don-t-get-initialized-properly-80c34018155646d08e4da0bc6c977ed9?pvs=4
   useEffect(() => {
     if (!isInitialised) {
-      setValue("");
+      setValue("0");
     }
   });
 
@@ -21,7 +22,7 @@ export function TextualBuilder(props: CadenceValueBuilder): ReactElement {
 
   return (
     <Input
-      type="text"
+      type="number"
       value={value}
       onChange={(e) => setValue(e.target.value)}
     />
