+++ frontend/src/modules/interactions/components/ValueBuilder/PathBuilder/PathBuilder.tsx
@@ -4,8 +4,8 @@ import { ReactElement, useEffect } from "react";
 import { CadenceValueBuilder } from "../interface";
 import SelectInput, {
   SelectInputOption,
-} from "../../../../../../components/inputs/select-input/SelectInput";
-import Input from "../../../../../../components/inputs/input/Input";
+} from "../../../../../components/inputs/select-input/SelectInput";
+import Input from "../../../../../components/inputs/input/Input";
 
 export function PathBuilder(props: CadenceValueBuilder): ReactElement {
   const { type, value, setValue } = props;
