+++ frontend/src/modules/interactions/components/ValueBuilder/TextualBuilder/TextualBuilder.tsx
@@ -1,18 +1,17 @@
 import React, { ReactElement, useEffect } from "react";
 import { CadenceValueBuilder } from "../interface";
-import Input from "../../../../../../components/inputs/input/Input";
-import { FclValues } from "@flowser/shared";
+import Input from "../../../../../components/inputs/input/Input";
 
-export function IntegerNumberBuilder(props: CadenceValueBuilder): ReactElement {
+export function TextualBuilder(props: CadenceValueBuilder): ReactElement {
   const { value, setValue } = props;
 
-  const isInitialised = FclValues.isFclIntegerNumberValue(value);
+  const isInitialised = typeof value === "string";
 
   // TODO(polish): Don't trigger this hook on every rerender
   //  See: https://www.notion.so/flowser/Sometimes-arguments-don-t-get-initialized-properly-80c34018155646d08e4da0bc6c977ed9?pvs=4
   useEffect(() => {
     if (!isInitialised) {
-      setValue("0");
+      setValue("");
     }
   });
 
@@ -22,7 +21,7 @@ export function IntegerNumberBuilder(props: CadenceValueBuilder): ReactElement {
 
   return (
     <Input
-      type="number"
+      type="text"
       value={value}
       onChange={(e) => setValue(e.target.value)}
     />
