+++ frontend/src/modules/interactions/components/ValueBuilder/ValueBuilder.tsx
@@ -9,10 +9,10 @@ import { AddressBuilder } from "./AddressBuilder/AddressBuilder";
 import { DictionaryBuilder } from "./DictionaryBuilder/DictionaryBuilder";
 import { PathBuilder } from "./PathBuilder/PathBuilder";
 import { BoolBuilder } from "./BoolBuilder/BoolBuilder";
-import { Callout } from "../../../../../components/callout/Callout";
-import { ExternalLink } from "../../../../../components/links/ExternalLink";
-import { FlowserIcon } from "../../../../../components/icons/Icons";
-import { SizedBox } from "../../../../../components/sized-box/SizedBox";
+import { Callout } from "../../../../components/callout/Callout";
+import { ExternalLink } from "../../../../components/links/ExternalLink";
+import { FlowserIcon } from "../../../../components/icons/Icons";
+import { SizedBox } from "../../../../components/sized-box/SizedBox";
 import { IntegerNumberBuilder } from "./IntegerNumberBuilder/IntegerNumberBuilder";
 
 export function ValueBuilder(props: CadenceValueBuilder): ReactElement {
