+++ frontend/src/modules/interactions/components/ValueBuilder/interface.ts
